Remake all three methods from QuestionResult and main refactor.

+++ src/main/java/kg/peaksoft/bilingualb4/api/payload/QuestionResultResponse.java
@@ -19,6 +19,6 @@ public class QuestionResultResponse {
     private String questionName;
     private int score;
     private Status status;
-    private int finalScore;
+    private double finalScore;
     private Status finalStatus;
 }
